[Print exit like tcsh]

+++ my_printf/my_putnbr_base_unsigned.c
@@ -13,7 +13,7 @@ int my_strlen(char const *str);
 
 int my_putnbr_base_unsigned(unsigned long long int nb, char const *base)
 {
-    int base_len = my_strlen(base);
+    unsigned long long int base_len = (unsigned long long int) my_strlen(base);
 
     if (nb >= base_len) {
         my_putnbr_base_unsigned(nb / base_len, base);
